+++ idb/postgres/internal/encoding/decode.go
@@ -8,14 +8,17 @@ import (
 	"github.com/algorand/go-algorand/data/bookkeeping"
 	"github.com/algorand/go-algorand/data/transactions"
 	"github.com/algorand/go-algorand/ledger/ledgercore"
-	"github.com/algorand/go-algorand/protocol"
+	"github.com/algorand/go-codec/codec"
 
 	"github.com/algorand/indexer/idb"
 	"github.com/algorand/indexer/idb/postgres/internal/types"
 )
 
 // DecodeJSON is a function that decodes json.
-var DecodeJSON = protocol.DecodeJSON
+func DecodeJSON(b []byte, objptr interface{}) error {
+	dec := codec.NewDecoderBytes(b, jsonCodecHandle)
+	return dec.Decode(objptr)
+}
 
 func decodeBase64(data string) ([]byte, error) {
 	return base64.StdEncoding.DecodeString(data)
