+++ idb/postgres/internal/encoding/encoding.go
@@ -1,0 +1,18 @@
+package encoding
+
+import (
+	"github.com/algorand/go-codec/codec"
+)
+
+var jsonCodecHandle *codec.JsonHandle
+
+func init() {
+	jsonCodecHandle = new(codec.JsonHandle)
+	jsonCodecHandle.ErrorIfNoField = true
+	jsonCodecHandle.ErrorIfNoArrayExpand = true
+	jsonCodecHandle.Canonical = true
+	jsonCodecHandle.RecursiveEmptyCheck = true
+	jsonCodecHandle.HTMLCharsAsIs = true
+	jsonCodecHandle.Indent = 0
+	jsonCodecHandle.MapKeyAsString = true
+}
